// packages/wizzi.backend/dist/src/features/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi-js\dist\lib\artifacts\ts\module\gen\main.js
    package: wizzi-js@0.7.9
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi\packages\wizzi.backend\.wizzi\src\features\types.ts.ittf
    utc time: Sun, 25 Jul 2021 19:40:41 GMT
*/

export type ValidateResult = { 
    isValid: boolean;
    message?: string;
};

export type CRUDResult = { 
    oper: string;
    ok: boolean;
    item?: any;
    message?: string;
};